requirements.txt
pyyaml
matplotlib
tqdm==4.66.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
scipy==1.12.0
opencv-python==4.9.0.80
imageio==2.34.0
h5py==3.10.0
hdf5storage==0.1.19
tifffile==2024.2.12
mat73==0.62
tensorboardX==2.6.2.2